requirements.txt
flask>=2.0.1
flask-pymongo>=2.3.0
python-dotenv>=0.19.0
flask-cors>=3.0.10
python-jose>=3.3.0
passlib>=1.7.4
pymongo>=4.3.3
requests>=2.31.0 